pyproject.toml
[tool.poetry]
name = "qlient"
version = "0.2.3-beta"
description = "A fast and modern graphql client designed with simplicity in mind."
authors = ["Daniel Seifert <info@danielseifert.ch>"]
maintainers = ["Daniel Seifert <info@danielseifert.ch>"]
license = "MIT"
readme = "README.md"
homepage = "https://qlient-org.github.io/python-qlient/"
repository = "https://github.com/qlient-org/python-qlient"
documentation = "https://qlient-org.github.io/python-qlient/"
keywords = ["graphql", "client", "api", "query", "mutation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
packages = [
    { include = "qlient", from = "src" }
]
include = [
    "docs/*"
]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
requests = "^2.27.1"
platformdirs = "^2.5.1"
importlib-metadata = "^4.11.4"
websocket-client = "^1.3.2"

[tool.poetry.dev-dependencies]
coverage = "^6.4.1"
pytest = "^7.0.1"
requests-mock = "^1.9.3"
strawberry-graphql = "^0.114.0"
hypothesis = "^6.37.2"
yappi = "^1.3.3"
mkdocs = "^1.2.3"
mkdocs-material = "^8.2.1"
mkdocs-exclude = "^1.0.2"
mkdocs-include-markdown-plugin = "^3.3.0"
black = "^22.3.0"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
